Double für Costs A2 Beispiel-Dateien

+++ MMI/A2/Program.cs
@@ -23,14 +23,14 @@ namespace A2
             g.AddVertex("5");
 
 
-            g.AddEdge(g.Vertices["1"], g.Vertices["2"], new Dictionary<string, int> { { "C", 13 } });
-            g.AddEdge(g.Vertices["1"], g.Vertices["3"], new Dictionary<string, int> { { "C", 4 } });
-            g.AddEdge(g.Vertices["1"], g.Vertices["5"], new Dictionary<string, int> { { "C", 9 } });
-            g.AddEdge(g.Vertices["3"], g.Vertices["2"], new Dictionary<string, int> { { "C", 18 } });
-            g.AddEdge(g.Vertices["3"], g.Vertices["4"], new Dictionary<string, int> { { "C", 7 } });
-            g.AddEdge(g.Vertices["3"], g.Vertices["5"], new Dictionary<string, int> { { "C", 10 } });
-            g.AddEdge(g.Vertices["2"], g.Vertices["4"], new Dictionary<string, int> { { "C", 12 } });
-            g.AddEdge(g.Vertices["4"], g.Vertices["5"], new Dictionary<string, int> { { "C", 5 } });
+            g.AddEdge(g.Vertices["1"], g.Vertices["2"], new Dictionary<string, double> { { "C", 13 } });
+            g.AddEdge(g.Vertices["1"], g.Vertices["3"], new Dictionary<string, double> { { "C", 4 } });
+            g.AddEdge(g.Vertices["1"], g.Vertices["5"], new Dictionary<string, double> { { "C", 9 } });
+            g.AddEdge(g.Vertices["3"], g.Vertices["2"], new Dictionary<string, double> { { "C", 18 } });
+            g.AddEdge(g.Vertices["3"], g.Vertices["4"], new Dictionary<string, double> { { "C", 7 } });
+            g.AddEdge(g.Vertices["3"], g.Vertices["5"], new Dictionary<string, double> { { "C", 10 } });
+            g.AddEdge(g.Vertices["2"], g.Vertices["4"], new Dictionary<string, double> { { "C", 12 } });
+            g.AddEdge(g.Vertices["4"], g.Vertices["5"], new Dictionary<string, double> { { "C", 5 } });
 
 
             var a = MST.Prim(g, "C");
